fix several issues in mac os build - icons/images were not added to bundle - hdf5 uses long long since 1.10.0: several IO routines did not work correctly because of runtime errors in hdf5

+++ Core/HDF5Writer.h
@@ -21,7 +21,7 @@ class iSegCore_API HDF5Writer
 {
 public:
 	typedef unsigned long long size_type;
-	typedef int hid_type;
+	typedef long long hid_type;
 	HDF5Writer();
 	~HDF5Writer();
